requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mcp==1.18.0
pydantic==2.11.7
requests==2.31.0
python-dotenv==1.0.0